+++ drawer.cpp
@@ -124,6 +124,8 @@ Drawer::Drawer() : mglDraw(),
     proximity_curve_correlation = 0;
     proximity_curve_integral = 0;
     proximity_curve_local = 0;
+    proximity_curve_relative = 0;
+    proximity_curve_average_relative = 0;
 }
 
 Drawer::~Drawer()