+++ drawerdp.h
@@ -27,6 +27,7 @@ protected:
     double proximity_shape_mark, proximity_range_mark;
     int f0min, f0max;
     int userf0min, userf0max;
+    double rt, ru;
 
     SimpleGraphData * simple_data;
 };