+++ graphsevalwindow.cpp
@@ -47,6 +47,8 @@ Drawer * GraphsEvalWindow::createNewDrawer(QString path)
     this->ui->totalC->setText(QString::number(this->drawer->proximity_curve_correlation));
     this->ui->totalI->setText(QString::number(this->drawer->proximity_curve_integral));
     this->ui->totalL->setText(QString::number(this->drawer->proximity_curve_local));
+    this->ui->totalA->setText(QString::number(this->drawer->proximity_curve_relative));
+    this->ui->totalAR->setText(QString::number(this->drawer->proximity_curve_average_relative));
     this->ui->totalPr->setText(QString::number(this->drawer->proximity_range));
 
     SPTK_SETTINGS * sptk_settings = SettingsDialog::getSPTKsettings();
@@ -63,6 +65,14 @@ Drawer * GraphsEvalWindow::createNewDrawer(QString path)
     case 2:
         this->ui->totalL->setStyleSheet("font-weight: bold");
         this->ui->totalLtitle->setStyleSheet("font-weight: bold");
+        break;
+    case 3:
+        this->ui->totalA->setStyleSheet("font-weight: bold");
+        this->ui->totalAtitle->setStyleSheet("font-weight: bold");
+        break;
+    case 4:
+        this->ui->totalAR->setStyleSheet("font-weight: bold");
+        this->ui->totalARtitle->setStyleSheet("font-weight: bold");
         break;
     default:
         break;