+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.4] - 2018-02-18
+
+### Changed
+- Changed range mark calculation
+
+### Added
+- New calculation type for F0 curves - relative
+- New calculation type for F0 curves - average relative
+
 ## [0.2.3] - 2018-02-10
 
 ### Changed